First stable release of the program, featuring a CLI implemented with the 'argparse' module

+++ organizer/core.py
@@ -3,9 +3,9 @@ from .config import EXTENSION_MAP
 from pathlib import Path 
 import shutil 
 
-def organize_by_extension(base_path):
+def organize_by_extension(base_path, simulate = False, verbose = False):
     base_path = Path(base_path)
-
+    
     if not base_path.exists() or not base_path.is_dir():
         raise ValueError(f'Caminho Inválido: {base_path}')
 
@@ -13,10 +13,14 @@ def organize_by_extension(base_path):
         if file.is_file():
             ext = file.suffix 
             category = getCategory(ext, EXTENSION_MAP)
-            target_folder = base_path / category 
-            EnsureFolder(target_folder)
+            target_folder = base_path / category
             target_path = target_folder / file.name 
-
-            if target_path.exists():
-                target_path = target_folder / f'{file.stem}_copy{file.suffix}'
-            shutil.move(str(file), str(target_path))
+            
+            if verbose:
+                print(f"[INFO] '{file.name}' -> '{category}/'")
+            
+            if not simulate:
+                EnsureFolder(target_folder)
+                if target_path.exists():
+                    target_path = target_folder / f'{file.stem}_copy{file.suffix}'
+                shutil.move(str(file), str(target_path))
